+++ NOOB/src/main/java/org/twspring/noob/Service/CoachingSessionService.java
@@ -14,16 +14,22 @@ public class CoachingSessionService {
 
     private final CoachingSessionRepository coachingSessionRepository;
 
+    // CRUD get all
     public List<CoachingSession> getAllCoachingSessions() {
         return coachingSessionRepository.findAll();
     }
 
+    // CRUD register
     public void addCoachingSession(CoachingSession coachingSession) {
         coachingSessionRepository.save(coachingSession);
     }
 
+    // CRUD update
     public void updateCoachingSession(Integer id, CoachingSession coachingSessionDetails) {
-        CoachingSession coachingSession = coachingSessionRepository.findById(id).orElseThrow(() -> new ApiException("Coaching session not found"));
+        CoachingSession coachingSession = coachingSessionRepository.findCoachingSessionById(id);
+        if (coachingSession == null) {
+            throw new ApiException("Coaching session not found");
+        }
         coachingSession.setScheduledTime(coachingSessionDetails.getScheduledTime());
         coachingSession.setSessionType(coachingSessionDetails.getSessionType());
         coachingSession.setFeedBack(coachingSessionDetails.getFeedBack());
@@ -32,13 +38,24 @@ public class CoachingSessionService {
         coachingSessionRepository.save(coachingSession);
     }
 
+    // CRUD delete
     public void deleteCoachingSession(Integer id) {
-        CoachingSession coachingSession = coachingSessionRepository.findById(id).orElseThrow(() -> new ApiException("Coaching session not found"));
+        CoachingSession coachingSession = coachingSessionRepository.findCoachingSessionById(id);
+        if (coachingSession == null) {
+            throw new ApiException("Coaching session not found");
+        }
         coachingSessionRepository.delete(coachingSession);
     }
 
+    // EXTRA ENDPOINT: getting a coaching session by id
     public CoachingSession getCoachingSessionById(Integer id) {
-        return coachingSessionRepository.findById(id).orElseThrow(() -> new ApiException("Coaching session not found"));
+        CoachingSession coachingSession = coachingSessionRepository.findCoachingSessionById(id);
+        if (coachingSession == null) {
+            throw new ApiException("Coaching session not found");
+        }
+        return coachingSession;
     }
+
+
 
 }
